feat: Implement ProductCategoryService and ProductService enhancements
- Added ProductCategoryService interface with methods for updating, deleting, and creating multiple product categories.
- Implemented ProductCategoryServiceImpl with validation, image handling, and asynchronous processing.
- Enhanced ProductService to support batch creation of products and product details, including image and video uploads.
- Introduced utility classes for calculating product hype and applying discounts.
- Updated GraphQL schema to reflect new product and category structures.
- Improved error handling and logging throughout service implementations.
- Configured application properties for GraphQL and file upload settings.

+++ src/main/java/com/v01/techgear_server/product/dto/ProductFilterSortResponse.java
@@ -2,6 +2,8 @@ package com.v01.techgear_server.product.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @AllArgsConstructor
@@ -9,7 +11,7 @@ import lombok.*;
 public class ProductFilterSortResponse {
     private Long id;
     private String name;
-    private double productPrice;
+    private BigDecimal productPrice;
     private double minPrice;
     private double maxPrice;
     private String category;
